requirements.txt
absl-py @ file:///C:/ci/absl-py_1623867344543/work
aiohttp @ file:///C:/ci/aiohttp_1614361212023/work
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
blinker==1.4
Bottleneck @ file:///C:/ci/bottleneck_1607557040328/work
brotlipy==0.7.0
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi @ file:///C:/ci/cffi_1625831756778/work
cftime @ file:///C:/ci/cftime_1638358653367/work
chardet @ file:///C:/ci/chardet_1605296577955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
coverage @ file:///C:/ci/coverage_1614615100481/work
cryptography @ file:///C:/ci/cryptography_1616769504165/work
Cython @ file:///C:/ci/cython_1626257072920/work
et-xmlfile==1.1.0
flatbuffers==1.12
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
GDAL==3.0.2
google-auth @ file:///tmp/build/80754af9/google-auth_1626320605116/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpustat==0.3.2
grpcio==1.34.1
h5py @ file:///C:/ci/h5py_1637120894255/work
idna @ file:///tmp/build/80754af9/idna_1622654382723/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877477759/work
jdcal @ file:///Users/ktietz/demo/mc3/conda-bld/jdcal_1630584345063/work
keras==2.9.0
keras-flops==0.1.2
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
Markdown @ file:///C:/ci/markdown_1614364082838/work
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multidict @ file:///C:/ci/multidict_1607349747897/work
netCDF4 @ file:///C:/ci/netcdf4_1624439444321/work
numexpr @ file:///C:/ci/numexpr_1618856728739/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1626271900803/work
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
olefile==0.46
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
openpyxl @ file:///tmp/build/80754af9/openpyxl_1615411699337/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
pandas @ file:///C:/ci/pandas_1644984307868/work
Pillow @ file:///C:/ci/pillow_1625663294916/work
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyhdf @ file:///C:/Users/neoooli/Documents/Downloads/pyhdf-0.10.5-cp39-cp39-win_amd64.whl
PyJWT @ file:///C:/ci/pyjwt_1619682721924/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy @ file:///C:/ci/scipy_1618856124610/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_baze37/tensorboard_1622025863084/work/tensorboard-2.5.0-py3-none-any.whl
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.5.0
tensorflow-estimator @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_baze37/tensorflow-estimator_1622026529081/work/tensorflow_estimator-2.5.0-py2.py3-none-any.whl
termcolor==1.1.0
torch-tb-profiler==0.4.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt @ file:///C:/ci/wrapt_1607574570428/work
xlrd @ file:///tmp/build/80754af9/xlrd_1608072521494/work
yarl @ file:///C:/ci/yarl_1606940155993/work
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
